% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aps.R
\name{estimate_aps}
\alias{estimate_aps}
\title{Main APS estimation function.}
\usage{
estimate_aps(
  data,
  ml,
  ml_type = "stats",
  xc = NA,
  xd = NA,
  infer = FALSE,
  s = 100,
  delta = 0.8,
  L = NA,
  seed = NA,
  fcn = NA,
  parallel = F,
  cores = NA,
  ...
)
}
\arguments{
\item{data}{Dataset containing ML input variables.}

\item{ml}{ML function for treatment recommendation}

\item{ml_type}{String indicating the ML object source package name, or "custom" for a user-defined
algorithm that runs prediction upon function call with a single input data.table object.
See details for supported packages. Defaults to "stats".}

\item{xc}{Character vector of column names of the continuous variables.}

\item{xd}{Character vector of column names of the discrete variables.}

\item{infer}{Boolean whether to infer continuous/discrete variables from remaining columns of data. Defaults to False.}

\item{s}{Number of draws for each APS estimation. Defaults to 100.}

\item{delta}{Radius of sampling ball. Can be either numeric or numeric vector. Defaults to 0.8.}

\item{L}{Named list where the names correspond to the names of the mixed
variables in the data, and the values are numeric vectors indicating the
set of discrete values for the variable.}

\item{seed}{Random seed}

\item{fcn}{Function to apply to output of ML function}

\item{parallel}{Boolean indicator for whether to parallelize the APS
estimation. Defaults to FALSE.}

\item{cores}{Integer number of cores for parallelization. If NA, then detectCores() is called.}

\item{\dots}{Additional inputs to be passed to \code{fcn}}
}
\value{
If a single \code{delta} value is passed, then the function returns a vector of
estimated Approximate Propensity Scores of the same length as the input data. If multiple
\code{delta} are passed, then a list of estimated APS vectors are returned, where the
keys are each \code{delta} value.
}
\description{
Main APS estimation function.
}
\details{
ML packages currently supported:
"mlr3", "caret", "stats" (base), "randomForest", "e1071", "bestridge", "rpart", "tree",
"custom" (for user-defined functions)

Approximate propensity score estimation involves taking draws \eqn{X_c^1,\ldots,X_c^s}
from the uniform distribution on
\eqn{N(X_{ci}, \delta)}, where \eqn{N(X_{ci},\delta)} is the
\eqn{p_c} dimensional ball centered at \eqn{X_{ci}} with radius \eqn{\delta}.
\eqn{X_c^1, \ldots,X_c^s} are destandardized before passed for ML
inference. The estimation equation is \eqn{p^s(X_i;\delta) = \frac{1}{s}\sum_{s=1}^{s} ML(X_c^s, X_{di})}.

If neither \code{xc} nor \code{xd} are passed and \code{infer=T}, then \code{data} is assumed to
be all relevant continuous inputs. If only one is passed, then the remaining
variables in \code{data} are assumed to be inputs of the other type. It is
recommended to pass both \code{xc} and \code{xd} in the case that not all the
variables in \code{data} are relevant for the ML input, instead of relying on inference.
}
\examples{
data("iris")
# Iris examples
model <- lm(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width, data=iris)
estimate_aps(iris, model, xc = names(iris)[2:3], xd = names(iris)[4],
            infer=FALSE, s=50, delta=0.1)

# Estimate APS while applying decision function assign_cutoff to model output with input cutoff=0.5
assign_cutoff <- function(X, cutoff){
  ret <- as.integer(X > cutoff)
  return(ret)
}
estimate_aps(iris, model, xc = names(iris)[2:3], xd = names(iris)[4],
             infer=FALSE, s=50, delta=0.1, fcn=assign_cutoff, cutoff=0.5)

# Multiple deltas
estimate_aps(iris, model, xc = names(iris)[2:3], xd = names(iris)[4],
             infer=FALSE, s=50, delta=c(0.1,0.5,1))

# Define mixed continuous/discrete variables
estimate_aps(iris, model, xc = names(iris)[2:3], xd = names(iris)[4],
             infer=FALSE, s=50, delta=0.1,
             L = list("Sepal.Width" = c(2, 3), "Petal.Length" = c(3, 4)))
}
