% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aps-package.R
\docType{package}
\name{IVaps}
\alias{IVaps}
\title{IVaps: Treatment effect estimation using Approximate Propensity Score from algorithmic assignment.}
\description{
The IVaps package offers functions for the two primary steps in this method of effect estimation:
Approximate Propensity Score estimation and instrumental variables estimation. The package also offers
a some diagnostic functions.
}
\section{IVaps functions}{


\code{\link{estimate_aps}}

\code{\link{estimate_treatment_effect}}

\code{\link{estimate_counterfactual_ml}}

\code{\link{covariate_balance_test}}
}

