% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qps-package.R
\docType{package}
\name{qps}
\alias{qps}
\title{qps: Treatment effect estimation using Quasi-Propensity Score from algorithmic assignment.}
\description{
The qps package offers functions for the two primary steps in this method of effect estimation:
Quasi-Propensity Score estimation and instrumental variables estimation. The package also offers
a some diagnostic functions.
}
\section{qps functions}{


\code{\link{estimate_qps}}

\code{\link{estimate_treatment_effect}}

\code{\link{estimate_counterfactual_ml}}

\code{\link{covariate_balance_test}}
}

