% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{covariate_balance_test}
\alias{covariate_balance_test}
\title{Covariate Balance Test}
\usage{
covariate_balance_test(data, qps = "QPS", X = c("X1"), Z = "Z", verbose = T)
}
\arguments{
\item{data}{Dataset containing ML input variables X, treatment assignment Z, and qps.}

\item{qps}{Column name of QPS variable. Defaults to "QPS".}

\item{X}{Character vector of column names of ML inputs X. Defaults to c("X1").}

\item{Z}{Column name of Z variable. Defaults to "Z".}

\item{verbose}{Boolean indicator for whether to print summary output of estimation. Defaults to True.}
}
\value{
List containing fitted multivariate multiple linear model and the joint hypothesis test.
}
\description{
Covariate Balance Test
}
\details{
This function estimates a multivariate multiple regression system, varying each of the ML input variables
in \code{X} against the common covariates \code{qps} and \code{Z}. The covariate balance test reports the
results of each individual regression, as well as the joint hypothesis result for the coefficient on \code{Z}.
This helps to establish whether \code{qps} performs adequately as a control for differences created through
treatment selection.
}
