% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qps.R
\name{estimate_qps}
\alias{estimate_qps}
\title{Main QPS estimation function.}
\usage{
estimate_qps(
  data,
  ML,
  ml_type = "stats",
  Xc = NA,
  Xd = NA,
  infer = F,
  S = 100,
  delta = 0.8,
  L = NA,
  seed = NA,
  fcn = NA,
  parallel = F,
  ...
)
}
\arguments{
\item{data}{Dataset containing ML input variables.}

\item{ML}{ML function for treatment recommendation}

\item{ml_type}{String indicating the ML object source package name, or "custom" for a user-defined
algorithm that runs prediction upon function call with a single input data.table object.
See details for supported packages. Defaults to "stats".}

\item{Xc}{Character vector of column names of the continuous variables.}

\item{Xd}{Character vector of column names of the discrete variables.}

\item{infer}{Boolean whether to infer continuous/discrete variables from remaining columns of data. Defaults to F.}

\item{S}{Number of draws for each QPS estimation. Defaults to 100.}

\item{delta}{Radius of sampling ball. Can be either numeric or numeric vector. Defaults to 0.8.}

\item{L}{Named list where the names correspond to the names of the mixed
variables in the data, and the values are numeric vectors indicating the
set of discrete values for the variable.}

\item{seed}{Random seed}

\item{fcn}{Function to apply to output of ML function}

\item{parallel}{Boolean indicator for whether to parallelize the QPS
estimation.}
}
\value{

}
\description{
Main QPS estimation function.
}
\details{
ML packages currently supported: mlr3, caret, stats (base)

Quasi-propensity score estimation involves taking draws \mjeqn{X_c^1,
\ldots,X_c^S}{X_c1, ..., X_cS} from the uniform distribution on
\mjeqn{N(X_{ci}, \delta)}, where \mjeqn{N(X_{ci},\delta)} is the \mjeqn{p_c}
dimensional ball centered at \mjeqn{X_{ci}} with radius \mjeqn{\delta}.
\mjeqn{X_c^1, \ldots,X_c^S} are destandardized before passed for ML
inference. The estimation equation is \mjeqn{p^s(X_i;\delta) = \frac{1}{S}
\sum_{s=1}^{S} ML(X_c^s, X_{di})}.

If neither \code{Xc} nor \code{Xd} are passed then \code{data} is assumed to
be all relevant continuous inputs. If only one is passed, then the remaining
variables in \code{data} are assumed to be inputs of the other type. It is
recommended to pass both \code{Xc} and \code{Xd} in the case that not all the
variables in \code{data} are relevant for the ML input.
}
