% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{estimate_treatment_effect}
\alias{estimate_treatment_effect}
\title{Main treatment effect estimation function}
\usage{
estimate_treatment_effect(
  data,
  qps = "QPS",
  Y = "Y",
  Z = "Z",
  D = "D",
  estimator = "2SLS",
  verbose = T
)
}
\arguments{
\item{data}{Dataset containing QPS, Y, Z, and D data.}

\item{qps}{Column name of QPS variable. Defaults to "QPS".}

\item{Y}{Column name of Y variable. Defaults to "Y".}

\item{Z}{Column name of Z variable. Defaults to "Z".}

\item{D}{Column name of D variable. Defaults to "D".}

\item{estimator}{String method of IV estimation. Defaults to "2SLS".}

\item{verbose}{Boolean indicator for whether to print summary output of estimation. Defaults to True.}
}
\value{
Fitted ivreg model object.
}
\description{
Main treatment effect estimation function
}
\details{
Treatment effect is estimated using IV estimation The default is to use the 2SLS method
of estimation, with the equations illustrated below.
\mjdeqn{D_i = \gamma_0(1-I) + \gamma_1 Z_i + \gamma_2 p^s(X_i;\delta) + v_i}{D_i = gamma_0(1-I) + gamma_1 Z_i + gamma_2 p^s(X_i;delta) + v_i}
\mjdeqn{Y_i = \beta_0(1-I) + \beta_1 D_i + \beta_2 p^s(X_i;\delta) + \epsilon_i}{Y_i = beta_0(1-I) + beta_1 D_i + beta_2 p^s(X_i;delta) + epsilon_i}
\mjeqn{\beta_1}{beta_1} is our causal estimation of the treatment effect. I is an indicator for if the ML funtion takes only a single
nondegenerate value in the sample.
}
\examples{

}
